Cleaned up code
Added Particle graph, cleaned up code. Updated Package.
