+++ Assets/Arduino/ByteReciever.cs.meta
@@ -1,0 +1,12 @@
+fileFormatVersion: 2
+guid: 396e6127ee9506a42a341ce6b280733d
+timeCreated: 1466699357
+licenseType: Pro
+MonoImporter:
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
