--- Assets/Arduino/LineMaterial.mat.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 3d55f7e8d1e712144ae4e67efb1e2f6b
-timeCreated: 1466555103
-licenseType: Pro
-NativeFormatImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
